Update carriers list
